pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hydros-agent-sdk"
version = "0.1.3"
description = "Official Python SDK for the Hydros ecosystem: Simulation Agents and MQTT Protocol"
readme = "README.md"
dependencies = [
    "paho-mqtt>=1.6.1",
    "pydantic>=2.0.0",
    "pyyaml>=6.0"
]
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "User Name", email = "weijiahao001@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
]

[project.scripts]
# hydros-sum = "hydros_sdk.cli:main"

[tool.setuptools.packages.find]
where = ["."]
